OAK-186: Avoid unnecessary rebase operations
git-svn-id: https://svn.apache.org/repos/asf/jackrabbit/oak/trunk/oak-core@1361266 13f79535-47bb-0310-9956-ffa450edef68

+++ src/main/java/org/apache/jackrabbit/oak/core/RootImpl.java
@@ -143,11 +143,13 @@ public class RootImpl implements Root {
 
     @Override
     public void rebase(ConflictHandler conflictHandler) {
-        purgePendingChanges();
-        NodeState base = getBaseState();
-        NodeState head = root.getNodeState();
-        refresh();
-        merge(base, head, root, conflictHandler);
+        if (!store.getRoot().equals(root.getBaseState())) {
+            purgePendingChanges();
+            NodeState base = getBaseState();
+            NodeState head = root.getNodeState();
+            refresh();
+            merge(base, head, root, conflictHandler);
+        }
     }
 
     @Override
